random fix on the event/post Manage
fetching the event json where the post is not archive
